research/02-pipeline-comparison/requirements.txt
# Food Detection Pipeline - Dependencies

# Object Detection (YOLO)
ultralytics>=8.1.0

# LLM Integration (OpenAI Vision)
openai>=1.12.0

# Image Processing
pillow>=10.2.0

# Environment Variables
python-dotenv>=1.0.0
